+++ djem/forms/__init__.py
@@ -1,2 +1,2 @@
-from djem.forms.bases import AuditableForm, CommonInfoForm, UserSaveMixin  # NOQA
+from djem.forms.bases import AuditableForm, CommonInfoForm, UserSavable  # NOQA
 from djem.forms.fields import TimeZoneField  # NOQA